Refactor validators into separate files
