+++ dftimewolf/lib/validators/__init__.py
@@ -1,0 +1,11 @@
+# -*- coding: utf-8 -*-
+"""This file imports Python modules that register validators."""
+from dftimewolf.lib.validators import aws_region
+from dftimewolf.lib.validators import azure_region
+from dftimewolf.lib.validators import datetime
+from dftimewolf.lib.validators import gcp_zone
+from dftimewolf.lib.validators import grr_host
+from dftimewolf.lib.validators import hostname
+from dftimewolf.lib.validators import regex
+from dftimewolf.lib.validators import subnet
+from dftimewolf.lib.validators import url
